requirements.txt
coloredlogs==15.0.1
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.9.0
humanfriendly==10.0
Jinja2==3.1.4
MarkupSafe==2.1.5
ml_dtypes==0.5.0
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
onnx==1.16.2
onnxruntime==1.19.2
onnxscript==0.1.0.dev20240930
packaging==24.1
protobuf==5.28.2
PyYAML==6.0.2
sympy==1.13.3
torch==2.4.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
